Refactor: use an interface instead of concrete class
